Rename scribereader module
