+++ mypy.ini
@@ -15,6 +15,3 @@ ignore_missing_imports = True
 
 [mypy-clog.*]
 ignore_missing_imports = True
-
-[mypy-scribereader.*]
-ignore_missing_imports = True
