+++ tox.ini
@@ -21,7 +21,7 @@ commands =
     pre-commit run --all-files
     # tron has been around for a while, so we'll need to slowly add types or make an effort
     # to get it mypy-clean in one shot - until then, let's only check files that we've added types to
-    mypy --pretty tron/kubernetes.py tron/commands/backfill.py bin/tronctl tron/utils/scribereader.py
+    mypy --pretty tron/kubernetes.py tron/commands/backfill.py bin/tronctl tron/utils/logreader.py
     check-requirements
     # optionally install yelpy requirements - this is after check-requirements since
     # check-requirements doesn't understand these extra requirements
